+++ app/Http/Controllers/Api/BatasBarangController.php
@@ -6,13 +6,18 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\BatasBarangResource;
 use App\Models\BatasBarang;
 use Illuminate\Http\Request;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Symfony\Component\HttpFoundation\Response as HttpResponse;
 
 class BatasBarangController extends Controller
 {
+    use AuthorizesRequests;
+
     // GET /api/batas-barang
     public function index()
     {
+        $this->authorize('viewAny', BatasBarang::class);
+
         $batas = BatasBarang::all();
         return BatasBarangResource::collection($batas)
             ->response()
@@ -22,6 +27,8 @@ class BatasBarangController extends Controller
     // POST /api/batas-barang
     public function store(Request $request)
     {
+        $this->authorize('create', BatasBarang::class);
+
         $data = $request->validate([
             'id_barang'   => 'required|string',
             'batas_barang'=> 'required|integer|min:0',
@@ -37,6 +44,8 @@ class BatasBarangController extends Controller
     public function show($id_barang)
     {
         $batas = BatasBarang::findOrFail($id_barang);
+        $this->authorize('view', $batas);
+
         return (new BatasBarangResource($batas))
             ->response()
             ->setStatusCode(HttpResponse::HTTP_OK);
@@ -46,6 +55,8 @@ class BatasBarangController extends Controller
     public function update(Request $request, $id_barang)
     {
         $batas = BatasBarang::findOrFail($id_barang);
+        $this->authorize('update', $batas);
+
         $data = $request->validate([
             'batas_barang'=> 'sometimes|required|integer|min:0',
         ]);
@@ -59,6 +70,8 @@ class BatasBarangController extends Controller
     public function destroy($id_barang)
     {
         $batas = BatasBarang::findOrFail($id_barang);
+        $this->authorize('delete', $batas);
+
         $batas->delete();
         return response()->json(null, HttpResponse::HTTP_NO_CONTENT);
     }
