+++ app/Http/Controllers/Api/JenisBarangController.php
@@ -7,11 +7,14 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\JenisBarangResource;
 use App\Models\JenisBarang;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response as HttpResponse;
 
 class JenisBarangController extends Controller
 {
+    use AuthorizesRequests;
+
     public function index()
     {
         return JenisBarangResource::collection(JenisBarang::all());
@@ -19,6 +22,8 @@ class JenisBarangController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize('create', JenisBarang::class);
+
         $data = $request->validate([
             'nama_jenis_barang' => 'required|string'
         ]);
@@ -30,12 +35,17 @@ class JenisBarangController extends Controller
     public function show($id)
     {
         $jenis = JenisBarang::findOrFail($id);
+
+        $this->authorize('view', $jenis);
+
         return new JenisBarangResource($jenis);
     }
 
     public function update(Request $request, $id)
     {
         $jenis = JenisBarang::findOrFail($id);
+
+        $this->authorize('update', $jenis);
 
         $data = $request->validate([
             'nama_jenis_barang' => 'required|string'
@@ -48,6 +58,9 @@ class JenisBarangController extends Controller
     public function destroy($id)
     {
         $jenis = JenisBarang::findOrFail($id);
+
+        $this->authorize('delete', $jenis);
+        
         $jenis->delete();
         return response()->json(null, HttpResponse::HTTP_NO_CONTENT);
     }
