+++ app/Http/Controllers/Api/PengajuanController.php
@@ -5,14 +5,18 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\PengajuanResource;
 use App\Models\Pengajuan;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response as HttpResponse;
 
 class PengajuanController extends Controller
 {
+    use AuthorizesRequests;
     // GET /api/pengajuan
     public function index()
     {
+        $this->authorize('viewAny', Pengajuan::class);
+
         $pengajuan = Pengajuan::with('user', 'details')->get();
         return PengajuanResource::collection($pengajuan);
     }
@@ -20,6 +24,8 @@ class PengajuanController extends Controller
     // POST /api/pengajuan
     public function store(Request $request)
     {
+        $this->authorize('create', Pengajuan::class);
+
         $data = $request->validate([
             'id_pengajuan'    => 'required|string',
             'unique_id'       => 'required|string',
@@ -27,7 +33,7 @@ class PengajuanController extends Controller
         ]);
 
         $pengajuan = Pengajuan::create($data);
-        return new PengajuanResource($pengajuan)
+        return (new PengajuanResource($pengajuan))
         ->response()
         ->setStatusCode(HttpResponse::HTTP_CREATED);
     }
@@ -38,7 +44,10 @@ class PengajuanController extends Controller
         $pengajuan = Pengajuan::with('user', 'details')
             ->where('id_pengajuan', $id_pengajuan)
             ->firstOrFail();
-        return new PengajuanResource($pengajuan)
+            
+        $this->authorize('view', $pengajuan);
+
+        return (new PengajuanResource($pengajuan))
             ->response()
             ->setStatusCode(HttpResponse::HTTP_OK);
     }
@@ -47,12 +56,15 @@ class PengajuanController extends Controller
     public function update(Request $request, $id_pengajuan)
     {
         $pengajuan = Pengajuan::where('id_pengajuan', $id_pengajuan)->firstOrFail();
+
+        $this->authorize('update', $pengajuan);
+
         $data = $request->validate([
             'status_pengajuan'=> 'sometimes|required|in:Menunggu Persetujuan,Disetujui,Ditolak',
         ]);
 
         $pengajuan->update($data);
-        return new PengajuanResource($pengajuan)
+        return (new PengajuanResource($pengajuan))
             ->response()
             ->setStatusCode(HttpResponse::HTTP_OK);
     }
@@ -61,6 +73,9 @@ class PengajuanController extends Controller
     public function destroy($id_pengajuan)
     {
         $pengajuan = Pengajuan::where('id_pengajuan', $id_pengajuan)->firstOrFail();
+
+        $this->authorize('delete', $pengajuan); 
+        
         $pengajuan->delete();
         return response()->json(null, HttpResponse::HTTP_NO_CONTENT);
     }
